pom - change order of default loading when root not specified.
Changed order of root pom injection when not specified to ensure
the configuration of the specified pom supercedes the root.
Remove foam's root pom envs:appName, as this makes it difficult
for application level poms to supercede, given how 'name' is
determined.
Update EnvMaker to delay setting APP_NAME as long as possible.